backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
openai==1.54.0
groq==1.0.0
python-dotenv==1.0.1
colormath==3.0.0
numpy==2.0.2
python-multipart==0.0.12
Pillow==11.0.0
scikit-learn==1.6.1
